LayoutTests/imported/w3c/web-platform-tests/encoding/legacy-mb-korean/euc-kr/euckr-decode-errors-expected.txt


PASS lead not 0x00 and no more bytes: B0 
PASS lead not 0x00 and no more bytes: B0 B5 B0 
PASS lead not 0x00 and no more bytes: B0 B0 B0 
FAIL lead byte outside 0x81-0xFE: FF assert_equals: expected "\ufffd" but got ""
FAIL lead byte outside 0x81-0xFE: B0 B0 FF assert_equals: expected "갠\ufffd" but got "갠"
PASS trail byte outside 0x41-0xFE: B0 31 
FAIL trail byte outside 0x41-0xFE: B0 FF assert_equals: expected "\ufffd" but got "\ufffd"
FAIL pointer is null: 81 5B assert_equals: expected "\ufffd[" but got "\ufffd"
